backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==19.12.0
